interaction: Use serial output instead of console one if possible
It's common for platforms with graphics cards to register the graphics
card as both console output device and graphics output protocol. In this
case, since the two output implementation actually shares the
framebuffer, the output produced by one may be overwritten by another,
which is undesired.

For console output with SIMPLE_TEXT_PROTOCOL, it's used to provide
output for devices with serial access only, which could now be replaced
by EFI_SERIAL_IO_OUTPUT protocol.

Console output now only serves as a fallback before serial is
initialized or in case that no serial devices could be probed.

Signed-off-by: Yao Zi <ziyao@disroot.org>