eficall: Fix stack alignment in ABI wrappers
Microsoft X64 ABI states,

> The stack will always be maintained 16-byte aligned, except within
> the prolog (for example, after the return address is pushed)

Since call pushes the return address (8 bytes) onto the stack on x86_64,
this means we should maintain the stack aligned to 16-bytes BEFORE issue
of the call, similar to SystemV ABI's case.

But in wrappers, an 8-byte return address is already pushed on the stack
by the caller, which means we must align the stack down to 16-bytes
manually, again.

Let's fix the wrappers, and this patch allows loli-loader to boot on
real-world x86_64 machines with AMI UEFI firmware.

Link: https://learn.microsoft.com/en-us/cpp/build/stack-usage?view=msvc-170&source=recommendations
Signed-off-by: Yao Zi <ziyao@disroot.org>