build: Ensure SIMD is disabled
UEFI specification is unclear about usage of SIMD registers.

For x86_64, XMM is explicitly permitted to be used (at least for passing
FP arguments), but YMM and ZMM aren't mentioned.

For RISC-V, UEFI only specifies,

	UEFI drivers will need to know the capabilities of processor
	before executing the specified RISC-V extension instructions.

so at least V extension need to be checked before usage. It's too
complex for a bootloader, so we want to avoid usage of V extension.

Unfortunately, there isn't something like -mno-general-registers or
-mno-lsx for RISC-V, and we don't want to override the -march argument
passed by the caller. Only a comment is added to raise packager's notice
of not passing inappropriate -march argument.

For LoongArch, UEFI states,

	FP unit can be used(CSR.EUEN.FPE to enable), calling convention
	refer to 2.3.8.2.

it's not sure whether SIMD instructions are considered as part of the FP
unit. Let's disable them with -mno-lsx and -mno-lasx for safety.

AArch64 is one of the only two architecture that clearly defines usage
of SIMD extensions (the other is AArch32),

	Floating point and SIMD instructions may be used.

But we still disallow usage of SIMD instructions with
"-mgeneral-regs-only" conservatively. It basically doesn't hurt for such
a small application.

Signed-off-by: Yao Zi <ziyao@disroot.org>