initrd: Pass initrd by EFI_LOAD_FILE protocol
The formal method to pass an initramfs to Linux kernel is by installing
a LoadFile2-capable handle with device path set to LINUX_INITRD_GUID.
Linux EFI stub will look for such a handle, and register an EFI
configuratin table with its content and LINUX_INITRD_GUID to be parsed
by the kernel latter.

Previously, loli-loader skipped the installation of the device handle,
instead a configuration table is installed and expected to be parsed by
kernel directly, which simplifies the code but actually isn't guaranteed
to work, since the configuration table isn't part of the ABI, but Linux
kernel's implementation details.

Let's correctly install a LoadFile2 handle to pass the initrd. It's a
little annoying to make our functions callable in Windows ABI. In hope
of this is the only place we need to do so, a wrapper is open-coded in
eficall.S.

Link: https;//github.com/torvalds/linux/commit/ec93fc371f01
Signed-off-by: Yao Zi <ziyao@disroot.org>